Fix confusion around Rails::Memcached raw parameter.
I got confused in pull #1 about the parameter to prevent marshal loading
when reading or writing to memcached.  ::Memcached takes a marshal
parameter, so I thought false needed to be passed in to get the raw value,
but actually Rails::Memcached switches the meaning for this parameter and
renames it raw.