Remove MemcachedSafety class
This class is a subclass of Memcached::Rails and the only behavior it
has is to not raise when a non-fatal exception happen.

The Memcached::Rails class was removed from memcached gem and since
the cache store have the same behavior we are removing MemcachedSafety
without replacement.

Users that want to have the same behavior can decorate the memcached
client instead inheriting from it and adding this behavior.