Bump rubygems version to >=3.3.13
Tests on rails edge are failing because of a recent change in rails
(https://github.com/rails/rails/pull/46817) that requires rubygems
to have a version of 3.3.13 or above. Because of this, bundler can't
complete installing dependencies and tests fail.

> rails-7.1.0.alpha requires rubygems version >= 3.3.13, which is
> incompatible with the current version, 3.2.33

Bumping the version of rubygems in CI following the documentation
(https://github.com/marketplace/actions/setup-ruby-jruby-and-truffleruby#rubygems)
allows all tests to complete succesfully.